+++ flows/group_test.go
@@ -16,7 +16,16 @@ import (
 )
 
 func TestGroupList(t *testing.T) {
+	env := envs.NewBuilder().Build()
+
 	source, err := static.NewSource([]byte(`{
+		"fields": [
+			{
+				"key": "gender",
+				"name": "Gender",
+				"type": "text"
+			}
+		],
 		"groups": [
 			{
 				"uuid": "e25852ea-b014-4ac1-9982-d6dcb0c2a1d5",
@@ -30,13 +39,21 @@ func TestGroupList(t *testing.T) {
 				"uuid": "f4f4b78e-f072-42e2-987d-f5c13da3166d",
 				"name": "Males",
 				"query": "gender = \"M\""
+			},
+			{
+				"uuid": "f4f4b78e-f072-42e2-987d-f5c13da3166d",
+				"name": "Broken",
+				"query": "xyz = \"X\""
 			}
 		]
 	}`))
 	require.NoError(t, err)
 
-	sa, err := engine.NewSessionAssets(source, nil)
+	sa, err := engine.NewSessionAssets(env, source, nil)
 	require.NoError(t, err)
+
+	// check we ignored broken group
+	assert.Equal(t, 3, len(sa.Groups().All()))
 
 	testers := sa.Groups().Get("990e1392-1f49-40c5-9662-f39609324bf9")
 	males := sa.Groups().Get("f4f4b78e-f072-42e2-987d-f5c13da3166d")
@@ -65,8 +82,6 @@ func TestGroupList(t *testing.T) {
 
 	assert.Equal(t, males, groups.FindByUUID("f4f4b78e-f072-42e2-987d-f5c13da3166d"))
 	assert.Nil(t, groups.FindByUUID("7cb12d0e-e163-492c-95b1-28549cd04fe4"))
-
-	env := envs.NewBuilder().Build()
 
 	// check use in expressions
 	test.AssertXEqual(t, types.NewXArray(testers.ToXValue(env), males.ToXValue(env)), groups.ToXValue(env))