+++ services/webhooks/service.go
@@ -12,22 +12,24 @@ import (
 type service struct {
 	httpClient     *http.Client
 	httpRetries    *httpx.RetryConfig
+	httpAccess     *httpx.AccessConfig
 	defaultHeaders map[string]string
 	maxBodyBytes   int
 }
 
 // NewServiceFactory creates a new webhook service factory
-func NewServiceFactory(httpClient *http.Client, httpRetries *httpx.RetryConfig, defaultHeaders map[string]string, maxBodyBytes int) engine.WebhookServiceFactory {
+func NewServiceFactory(httpClient *http.Client, httpRetries *httpx.RetryConfig, httpAccess *httpx.AccessConfig, defaultHeaders map[string]string, maxBodyBytes int) engine.WebhookServiceFactory {
 	return func(flows.Session) (flows.WebhookService, error) {
-		return NewService(httpClient, httpRetries, defaultHeaders, maxBodyBytes), nil
+		return NewService(httpClient, httpRetries, httpAccess, defaultHeaders, maxBodyBytes), nil
 	}
 }
 
 // NewService creates a new default webhook service
-func NewService(httpClient *http.Client, httpRetries *httpx.RetryConfig, defaultHeaders map[string]string, maxBodyBytes int) flows.WebhookService {
+func NewService(httpClient *http.Client, httpRetries *httpx.RetryConfig, httpAccess *httpx.AccessConfig, defaultHeaders map[string]string, maxBodyBytes int) flows.WebhookService {
 	return &service{
 		httpClient:     httpClient,
 		httpRetries:    httpRetries,
+		httpAccess:     httpAccess,
 		defaultHeaders: defaultHeaders,
 		maxBodyBytes:   maxBodyBytes,
 	}
@@ -41,7 +43,7 @@ func (s *service) Call(session flows.Session, request *http.Request) (*flows.Web
 		}
 	}
 
-	trace, err := httpx.DoTrace(s.httpClient, request, s.httpRetries, s.maxBodyBytes)
+	trace, err := httpx.DoTrace(s.httpClient, request, s.httpRetries, s.httpAccess, s.maxBodyBytes)
 	if trace != nil {
 		call := &flows.WebhookCall{Trace: trace}